move message id generator to cli/api
this will be useful in cli package for
proper serializing / deserializing
