+++ packages/cli/src/api/generateMessageId.test.ts
