+++ packages/cli/src/api/index.ts
@@ -2,4 +2,5 @@ import getFormat from "./formats"
 import { getCatalogs, getCatalogForFile } from "./catalog"
 
 export { createCompiledCatalog } from "./compile"
+export { generateMessageId } from "./generateMessageId"
 export { getFormat, getCatalogs, getCatalogForFile }
